refactor: structure and fonts
