+++ src/components/Card.tsx
@@ -1,7 +1,5 @@
 import { Component } from 'react';
 
-import './card.scss';
-
 class Card extends Component {
   render() {
     return (
