+++ src/components/Searh.tsx
@@ -1,7 +1,5 @@
 import { Component } from 'react';
 
-import './search.scss';
-
 class Search extends Component {
   render() {
     return (
