+++ src/index.css
@@ -1,3 +1,3 @@
 @import './styles/normalize.css';
 @import './styles/fonts.css';
-@import './styles/settings/index.scss';
+@import './styles/components/index.scss';
