+++ src/styles/components/_search.scss
@@ -1,0 +1,29 @@
+.search {
+  margin: 40px;
+
+  &__block {
+    @include flexCenterBetween;
+    column-gap: 48px;
+  }
+
+  input {
+    border: 2px solid $black;
+    border-radius: 12px;
+    width: 100%;
+    padding: 12px;
+    background-color: $grey;
+    color: $black;
+
+    &::placeholder {
+      color: $black;
+    }
+  }
+
+  button  {
+    border-radius: 12px;
+    padding: 16px 32px;
+    border: 2px solid $black;
+    background-color: $grey;
+
+  }
+}
