+++ src/styles/components/index.scss
@@ -1,0 +1,3 @@
+@import '../settings/index.scss';
+
+@import './search.scss';
